requirements.txt
# ---------------------------------------------
# 📊 DATA PROCESSING & ANALYSIS
# ---------------------------------------------
pandas==1.5.3          # Data manipulation and analysis — core for working with tabular data
numpy==1.24.3          # Numerical operations, arrays, and matrix support (used by almost all ML libraries)

# ---------------------------------------------
# 🧠 MACHINE LEARNING
# ---------------------------------------------
scikit-learn==1.2.2    # Classic ML models (regression, classification, clustering), preprocessing, model evaluation
xgboost==1.7.5
# ---------------------------------------------
# 📈 VISUALIZATION
# ---------------------------------------------
matplotlib==3.7.1      # Base library for plotting graphs, charts, and visualizations
seaborn==0.12.2        # High-level interface for drawing attractive and informative statistical graphics (built on matplotlib)

# ---------------------------------------------
# 📦 EXPERIMENT TRACKING & MODEL MANAGEMENT
# ---------------------------------------------
mlflow==2.3.1          # End-to-end ML lifecycle tracking, including experiment tracking, model packaging, and deployment

# ---------------------------------------------
# ✅ TESTING
# ---------------------------------------------
pytest==7.3.1          # Python testing framework — great for writing unit tests for ML pipelines, data validation, etc.

# ---------------------------------------------
# ⚡ API DEVELOPMENT (FOR MODEL SERVING)
# ---------------------------------------------
fastapi==0.95.2        # Lightweight, high-performance web framework for serving ML models via REST APIs
uvicorn==0.22.0        # ASGI server for running FastAPI apps — lightweight and fast
#
# ---------------------------------------------
#  MISC
# ---------------------------------------------
pyyaml==6.0
joblib==1.3.1
setuptools==65.5.0
ipykernel==6.29.5


# ---------------------------------------------
# 🔧 PIPELINE & AUTOMATION
# ---------------------------------------------
scipy==1.10.1           # Scientific computing library for statistical tests
requests==2.31.0       # HTTP library for API calls and health checks
streamlit==1.24.0      # Web app framework for ML dashboards
plotly==5.14.1         # Interactive plotting library
pydantic==1.10.8       # Data validation and settings management